+++ field_utils/FieldCondensedArray.F90
@@ -1,7 +1,7 @@
 #include "MAPL_Generic.h"
 module mapl3g_FieldCondensedArray
    use, intrinsic :: iso_c_binding, only: c_ptr, c_f_pointer
-   use mapl3g_FieldCondensedArray_private, only: ARRAY_RANK, get_shape => get_fptr_shape
+   use mapl3g_FieldCondensedArray_private, only: ARRAY_RANK, get_fptr_shape_private
    use mapl3g_FieldDimensionInfo, only: get_vertical_dim_spec_name
    use MAPL_FieldPointerUtilities, only: FieldGetLocalElementCount, assign_fptr
    use MAPL_ExceptionHandling
@@ -56,16 +56,20 @@ contains
       logical :: has_vertical
       character(len=:), allocatable :: spec_name
       character(len=*), parameter :: VERTICAL_DIM_NONE_NAME = 'VERTICAL_DIM_NONE'
+      integer :: dimCount
 
-      call ESMF_FieldGet(f, gridToFieldMap=gridToFieldMap, _RC) 
-      call ESMF_FieldGet(f, rank=rank, _RC)
+      call ESMF_FieldGet(f, dimCount=dimCount, rank=rank, _RC)
+      _ASSERT(.not. rank < 0, 'rank cannot be negative.')
+      _ASSERT(.not. dimCount < 0, 'dimCount cannot be negative.')
       allocate(localElementCount(rank))
+      allocate(gridToFieldMap(dimCount))
+      call ESMF_FieldGet(f, gridToFieldMap=gridToFieldMap, _RC) 
       !  Due to an ESMF bug, getting the localElementCount must use the module function.
       !  See FieldGetLocalElementCount (specific function) comments.
       localElementCount = FieldGetLocalElementCount(f, _RC)
       spec_name = get_vertical_dim_spec_name(f, _RC)
       has_vertical = spec_name /= VERTICAL_DIM_NONE_NAME
-      fptr_shape = get_shape(gridToFieldMap, localElementCount, has_vertical, _RC)
+      fptr_shape = get_fptr_shape_private(gridToFieldMap, localElementCount, has_vertical, _RC)
 
    end function get_fptr_shape